Use namespacing for view components
