+++ app/helpers/view_components/maintaining_iterations_summary_table.rb
@@ -1,0 +1,13 @@
+module ViewComponents
+  class MaintainingIterationsSummaryTable < ViewComponent
+    include Mandate
+
+    initialize_with :iterations
+
+    def to_s
+      react_component("maintaining-iterations-summary-table", {
+                        iterations: iterations.map(&:serialized)
+                      })
+    end
+  end
+end
