+++ app/helpers/view_components/mentor_inbox.rb
@@ -1,0 +1,15 @@
+module ViewComponents
+  class MentorInbox < ViewComponent
+    include Mandate
+
+    initialize_with :conversations_request, :tracks_request, :sort_options
+
+    def to_s
+      react_component("mentor-inbox", {
+                        conversations_request: conversations_request,
+                        tracks_request: tracks_request,
+                        sort_options: sort_options
+                      })
+    end
+  end
+end
