--- app/view_components/view_component.rb
@@ -1,10 +1,0 @@
-class ViewComponent
-  include ActionView::Helpers::TagHelper
-
-  def react_component(id, data)
-    tag :div, {
-      "data-react-#{id}": true,
-      "data-react-data": data.to_json
-    }
-  end
-end
