+++ app/views/maintaining/iterations/index.html.haml
@@ -104,4 +104,4 @@
         = button_tag "Add track"
   .info
     %h2 Iterations
-    = MaintainingIterationsSummaryTable.new(@iterations)
+    = ViewComponents::MaintainingIterationsSummaryTable.new(@iterations)
