+++ app/views/test/components/example/iterations_summary_table/index.html.haml
@@ -1,1 +1,1 @@
-= ExampleIterationsSummaryTable.new(@solution)
+= ViewComponents::ExampleIterationsSummaryTable.new(@solution)
