+++ app/views/test/components/maintaining/iterations_summary_table/index.html.haml
@@ -1,1 +1,1 @@
-= MaintainingIterationsSummaryTable.new(@iterations)
+= ViewComponents::MaintainingIterationsSummaryTable.new(@iterations)
