+++ app/views/test/components/notifications/icons/show.html.haml
@@ -3,4 +3,4 @@
   = f.label :count
   = f.number_field :count
   = f.submit "Update"
-= NotificationIcon.new(@user)
+= ViewComponents::NotificationIcon.new(@user)
